added serverless function for vercel deployment
