+++ server/vercel.json
@@ -1,1 +1,1 @@
-{"rewrites": [{ "source": "/(.*)", "destination": "/app.js" }] }
+{"rewrites": [{ "source": "/(.*)", "destination": "/api" }] }
